[MODULES-372] provide our own "java.se" module on JDK 9+
The `java.se` module in JDK 9+ isn't "available" by default. However,
it's merely an aggregator of other modules that _are_ "available".
This commit provides our own `java.se` module even on JDK 9+, instead
of delegating to JPMS. The new `java.se` module aggregates the same
modules as the `java.se` module in the JDK, the only difference is
that it's constructed on the JBoss Modules level and hence doesn't
require users to start JVM with `--add-modules java.se`.